requirements.txt
annotated-types==0.6.0
coverage==7.4.4
et-xmlfile==1.1.0
Faker==24.3.0
iniconfig==2.0.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.1
pluggy==1.4.0
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==4.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
typing_extensions==4.10.0
tzdata==2024.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
yattag==1.15.2
